gjorde om profilepage och userprofile komponenten så att rätt kod låg i rätt fil.
